+++ R/documents.R
@@ -127,7 +127,8 @@ documents <- function(..., limit = NULL) {
       resp_pages = \(resp) resp_total_pages(resp, per_page),
       resp_complete = \(resp) length(resp_documents(resp)) == 0L
     ),
-    max_reqs = max_reqs
+    max_reqs = max_reqs,
+    progress = wb_progress()
   )
 
   data <- resps_data(resps, resp_documents)